pyproject.toml
[tool.poetry]
name = "cdd-claude"
version = "0.1.2"
description = "AI-first development framework for Claude Code - conversational specs, autonomous planning, context-aware implementation"
authors = ["Guilherme Gouw <guilherme.gouw@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/guilhermegouw/context-driven-documentation"
repository = "https://github.com/guilhermegouw/context-driven-documentation"
keywords = ["ai", "claude", "documentation", "dev-tools", "llm", "ai-assisted-development", "requirements", "specifications"]
packages = [{include = "cddoc", from = "src"}]
include = [
    "src/cddoc/commands/*.md",
    "src/cddoc/templates/*.md",
    "src/cddoc/templates/*.yaml"
]

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Documentation",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Software Development :: Quality Assurance",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
    "Environment :: Console",
    "Framework :: Pytest",
]

[tool.poetry.dependencies]
python = "^3.9"
pyyaml = "^6.0.1"
click = "^8.1.7"
rich = "^13.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
black = "^23.12.1"
ruff = "^0.1.9"

[tool.poetry.scripts]
cdd = "cddoc.cli:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'

[tool.ruff]
line-length = 79
target-version = "py39"
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=cddoc --cov-report=term-missing"
